Added support for processing multiple tag sets at a time.
Added support to allow the user to specify multiple tag sets to download posts from, such that the user can download posts from e621 for multiple searches without having to update tags.txt and re-run the script between each set of tags. The convention for doing this is to specify each distinct set of tags to search against on its own line in tags.txt.